+++ spec/node_spec.rb
@@ -20,9 +20,14 @@
 # https://github.com/alces-software/metalware
 #==============================================================================
 
+require 'spec_helper'
+
 require 'node'
 require 'spec_utils'
-
+require 'fileutils'
+require 'config'
+require 'constants'
+require 'fakefs_helper'
 
 RSpec.describe Metalware::Node do
   before do
@@ -67,6 +72,51 @@ RSpec.describe Metalware::Node do
           }
         }
       })
+    end
+  end
+
+  describe "#answers" do
+    it 'performs a deep merge of answer files' do
+      config = Metalware::Config.new
+      fshelper = FakeFSHelper.new(config)
+      answers = Dir[File.join(FIXTURES_PATH, "answers/node-test-set1/*")]
+      fshelper.load_config_files
+      fshelper.add_answer_files(answers)
+      expected_hash = {
+        value_set_by_domain: "domain",
+        value_set_by_ag1: "ag1",
+        value_set_by_ag2: "ag2",
+        value_set_by_answer1: "answer1"
+      }
+
+      result_hash = fshelper.run do
+        Metalware::Node.new(config, 'answer1').answers
+      end
+      expect(result_hash).to eq(expected_hash)
+    end
+
+    # XXX factor out duplication between this and above
+    it 'just includes domain answers for nil node name' do
+      config = Metalware::Config.new
+      fshelper = FakeFSHelper.new(config)
+      answers = Dir[File.join(FIXTURES_PATH, "answers/node-test-set1/*")]
+      fshelper.load_config_files
+      fshelper.add_answer_files(answers)
+
+      # A nil node uses no configs but the 'domain' config, so all answers will
+      # be loaded from the 'domain' answers file.
+      expected_hash = {
+        value_set_by_domain: "domain",
+        value_set_by_ag1: "domain",
+        value_set_by_ag2: "domain",
+        value_set_by_answer1: "domain"
+      }
+
+      result_hash = fshelper.run do
+        node = Metalware::Node.new(config, nil)
+        node.answers
+      end
+      expect(result_hash).to eq(expected_hash)
     end
   end